Added Individual Project Sections
I added an animation and project sections for each of my two projects. currently they are simply sliding the information in on the x axis, and holding the content off screen, then hiding it on the way out. I will come back to this later and find a more elegant solution, but this works for now.