Finalized Rough Draft
Finalized the rough draft of the site, I added a color pallet as well as a sticky nav. I also went though and designed the site for a 1920 width screen. There may still be some changes to come with design, but form here I can work on getting the site live.